+++ ui/pages/asset/utils/security-trust-utils.test.ts
@@ -1,6 +1,5 @@
 import { SolScope } from '@metamask/keyring-api';
 import {
-  getSecurityTrustBlockExplorerLink,
   getSecurityTrustTokenTypeLabel,
   toSecurityTrustChainId,
 } from './security-trust-utils';
@@ -44,68 +43,6 @@ describe('security-trust-utils', () => {
           false,
         ),
       ).toBe('SPL');
-    });
-  });
-
-  describe('getSecurityTrustBlockExplorerLink', () => {
-    it('returns an EVM explorer link', () => {
-      const link = getSecurityTrustBlockExplorerLink({
-        caipChainId: 'eip155:1',
-        tokenAddress: '0xabc',
-        isNative: false,
-        evmNetworkConfigurations: {
-          '0x1': {
-            chainId: '0x1',
-            name: 'Ethereum Mainnet',
-            defaultBlockExplorerUrlIndex: 0,
-            blockExplorerUrls: ['https://etherscan.io'],
-            rpcEndpoints: [],
-            defaultRpcEndpointIndex: 0,
-          },
-        },
-        multichainNetworkConfigurations: {
-          'eip155:1': {
-            chainId: 'eip155:1',
-            name: 'Ethereum Mainnet',
-          },
-        },
-        fallbackExplorerLabel: 'Etherscan',
-      });
-
-      expect(link?.url).toContain('etherscan.io');
-      expect(link?.name).toBe('Ethereum Mainnet');
-    });
-
-    it('returns a non-EVM explorer link', () => {
-      const link = getSecurityTrustBlockExplorerLink({
-        caipChainId: SOLANA_MAINNET,
-        tokenAddress: 'So11111111111111111111111111111111111111112',
-        isNative: false,
-        evmNetworkConfigurations: {},
-        multichainNetworkConfigurations: {
-          [SOLANA_MAINNET]: {
-            chainId: SOLANA_MAINNET,
-            name: 'Solana',
-          },
-        },
-        fallbackExplorerLabel: 'Etherscan',
-      });
-
-      expect(link?.url).toContain('solscan.io');
-      expect(link?.name).toBeTruthy();
-    });
-
-    it('returns null for native assets', () => {
-      expect(
-        getSecurityTrustBlockExplorerLink({
-          caipChainId: 'eip155:1',
-          tokenAddress: '0xabc',
-          isNative: true,
-          evmNetworkConfigurations: {},
-          multichainNetworkConfigurations: {},
-          fallbackExplorerLabel: 'Etherscan',
-        }),
-      ).toBeNull();
     });
   });
 });
